requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
cffi==1.15.1
channels==4.0.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-money==3.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.6
packaging==23.2
Pillow==10.0.0
psycopg2==2.9.9
py-moneyed==3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
zope.interface==6.0
